KP-6162 Refactor `File`s: introduce subclasses for file types
Introduced a framework for having separate subclasses for different
types of files. No file type detection yet: the factory method for
creating `File`s just assigns everything into `UnknownTypeFile`, but
this is to change in KP-6161.

Pros of this design:
* avoiding if-elif-elif... copypaste in multiple methods: if we add a
  new datatype, we just create a new class and populate all relevant
  methods, as opposed to creating a new elif to all methods in the
  original `File` class
* keeping the (possibly messy) logic of determinign the file type
  separate from other business logic
* easy custom implementations for fixity checking, download urls etc
  based on file type
* getting rid of the Enum
